requirements.txt
Django==5.2.3
djangorestframework==3.16.0
python-dotenv==1.1.1
pillow==11.3.0
django-imagekit==5.0.0
djangorestframework_simplejwt==5.5.0

pytest-django==4.11.1
flake8==7.3.0
isort==6.0.1